fix and development authentication process

+++ database/migrations/2025_06_13_100338_create_merchant_types_table.php
@@ -11,7 +11,7 @@ return new class () extends Migration {
     public function up(): void
     {
         Schema::create('merchant_types', function (Blueprint $table) {
-            $table->tinyInteger('id')->unsigned()->autoIncrement();
+            $table->integer('id')->unsigned()->autoIncrement();
             $table->string('name', 100);
             $table->string('code', 50);
             $table->json('default_settings')->nullable();
